+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 0.4.0 (2021-01-12)
+
+[compare](https://github.com/metosin/reagent-dev-tools/compare/0.3.0...0.3.1)
+
+- Add way to control collection description text for custom types,
+like Linked:
+
+```
+(state-tree/register-collection-info-handler
+  lm/LinkedMap
+  #(state-tree/collection-info-handler "LinkedMap" "{LinkedMap, " (count %) " keys}"))
+```
+
 # 0.3.1 (2020-12-30)
 
 [compare](https://github.com/metosin/reagent-dev-tools/compare/0.3.0...0.3.1)